+++ src/pages/AddObservation/AddObservation.tsx
@@ -1241,6 +1241,7 @@ export default function AddObservation() {
               action={buttonClicked}
               disabled={addButtonDisabled()}
               primary
+              testId="addObservationButton"
               title={'button.add'}
             />
           </Grid>