+++ src/services/axios/sensitivityCalculator/getSensitivityCalculatorAPIData.ts
@@ -3,6 +3,7 @@ import GetWeighting from './getWeighting/getWeighting';
 import { helpers } from '../../../utils/helpers';
 import Observation from '../../../utils/types/observation';
 import Target from '../../../utils/types/target';
+import { SensCalcResults } from '../../../utils/types/sensCalcResults';
 import {
   TYPE_ZOOM,
   STATUS_PARTIAL,
@@ -11,9 +12,8 @@ import {
 } from '../../../utils/constants';
 import calculateSensitivityCalculatorResults from './calculateSensitivityCalculatorResults';
 import { SENSCALC_CONTINUUM_MOCKED } from '../../axios/sensitivityCalculator/SensCalcResultsMOCK';
-import { SensCalcResult } from '../../../utils/types/result';
 
-const SENSCALC_ERROR: SensCalcResult = {
+const SENSCALC_ERROR: SensCalcResults = {
   title: '',
   statusGUI: STATUS_ERROR,
   error: '',
@@ -22,11 +22,11 @@ const SENSCALC_ERROR: SensCalcResult = {
   section3: []
 };
 
-export const SENSCALC_LOADING: SensCalcResult = {
+export const SENSCALC_LOADING: SensCalcResults = {
   statusGUI: STATUS_PARTIAL
 };
 
-async function getSensCalc(observation: Observation, target: Target): Promise<SensCalcResult> {
+async function getSensCalc(observation: Observation, target: Target): Promise<SensCalcResults> {
   if (USE_LOCAL_DATA_SENSITIVITY_CALC) {
     return Promise.resolve(SENSCALC_CONTINUUM_MOCKED);
   }
@@ -75,7 +75,7 @@ async function getSensCalc(observation: Observation, target: Target): Promise<Se
       title: target.name,
       status: STATUS_ERROR
     });
-    return results as SensCalcResult;
+    return results as SensCalcResults;
   }
 }