+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Future
 * STAR-469 : Upgraded Observation page so it can support multiple target/observation combinations
 * STAR-537 : Added the ability to Edit a target
 * STAR-547 : Migrate page validations into separate utility 
+* STAR-573 : Add e2e Testing into the CI/CD Pipeline
+* STAR-575 : Update target list on target page to show Redshift & Velocity
 * STAR-587
    - Migrated Image Weighting to a separate component with testing coverage of 100%
    - Updated the Observation selection to remove duplicates and update fields correctly when selected
@@ -34,13 +36,13 @@ Future
    - Tapering field should not be available on Low.
    - Weather field should not be available for Low, only Mid
 * STAR-469 : Upgraded Observation page so it can support multiple target/observation combinations
-* Observation page: 
-   - STAR-508: Updated spectral resolution and effective resolution according to bandwidth, array selection and tapering values for Zoom modes
+* STAR-508: Updated spectral resolution and effective resolution according to bandwidth, array selection and tapering values for Zoom modes
 * STAR-540 : Proposal mapping:
    - Update getProposalList mapping to new backend proposal format
    - Move front-end to backend mapping into putProposal and postProposal
    - Update postProposal mapping
    - Update proposal Backend types and Mocks
+   - Upgrade to version 2.0.1 of SKA-PHT-SERVICES and 5.2.0 of SKA-DB-ODA
 * STAR-529: Implement validation from OSD with validate endpoint
 
 0.3.1